feat(parser): refactored background service
