+++ src/Services/Parser/BackgroundTasks/ParserService/ParserService.Core.Application/GlobalUsing.cs
@@ -1,4 +1,8 @@
 ﻿global using Common.Models;
 global using Repository.Base;
-global using ParserService.Core.Domain.Models;
 global using EventBus.RabbitMq.Interfaces;
+global using ParserService.Core.Domain.Models;
+global using ParserService.Core.Application.Handlers;
+global using ParserService.Core.Application.Services;
+global using ParserService.Core.Application.Interfaces;
+global using ParserService.Core.Application.Interfaces.Repositories;
