+++ src/Services/Parser/BackgroundTasks/ParserService/ParserService.Core.Application/LayerRegistration.cs
@@ -1,6 +1,4 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using ParserService.Core.Application.Interfaces;
-using ParserService.Core.Application.Services;
 
 namespace ParserService.Core.Application
 {
@@ -8,7 +6,7 @@ namespace ParserService.Core.Application
     {
         public static IServiceCollection AddApplicationLayer(this IServiceCollection services)
             => services
-                .AddSingleton<IAdHandlerService, AdHandlerService>()
+                .AddSingleton<IAdHandler, AdHandler>()
                 .AddSingleton<IParserBackgroundService, ParserBackgroundService>();
     }
 }
