+++ src/Services/Parser/BackgroundTasks/ParserService/ParserService.Core.Application/Services/ParserBackgroundService.cs
@@ -1,17 +1,15 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using ParserService.Core.Application.Interfaces;
-using ParserService.Core.Application.Interfaces.Repositories;
 
 namespace ParserService.Core.Application.Services
 {
     public class ParserBackgroundService : IParserBackgroundService
     {
         private readonly IServiceProvider _serviceProvider;
-        private readonly IAdHandlerService _adHandlerService;
+        private readonly IAdHandler _adHandlerService;
 
         public ParserBackgroundService(
             IServiceProvider serviceProvider,
-            IAdHandlerService adHandlerService)
+            IAdHandler adHandlerService)
         {
             _adHandlerService = adHandlerService;
             _serviceProvider = serviceProvider;
@@ -19,7 +17,9 @@ namespace ParserService.Core.Application.Services
 
         public async Task ExecuteAsync(CancellationToken cancellationToken)
         {
-            var siteDescriptions = await GetSiteDescriptions();
+            using var scope = _serviceProvider.CreateScope();
+            var repository = scope.ServiceProvider.GetRequiredService<ISiteDescriptionRepositoryAsync>();
+            var siteDescriptions = await repository.GetAllAsync(cancellationToken);
 
             Parallel
                 .ForEach(
@@ -28,15 +28,10 @@ namespace ParserService.Core.Application.Services
                     {
                         CancellationToken = cancellationToken
                     },
-                    body: siteDescription
-                    => _adHandlerService.AdHandlerAsync(siteDescription, cancellationToken));
-        }
-
-        private async Task<IEnumerable<SiteDescription>> GetSiteDescriptions()
-        {
-            using var scope = _serviceProvider.CreateScope();
-            var repository = scope.ServiceProvider.GetRequiredService<ISiteDescriptionRepositoryAsync>();
-            return await repository.GetAllAsync();
+                    body: async (siteDescription) => 
+                        await _adHandlerService.HandleAsync(
+                            siteDescription: siteDescription, 
+                            cancellationToken: cancellationToken));
         }
     }
 }
